Revert to group_menu_form_node_form_alter to provide group menus on the node form
Editing the node type form in c7bf212b1a89bdf001352465cc2fde9a7c05d7cb wasn't a satisfactory solution because that sets the available menus to those that the user who saves the form has access to. The node form does need to be altered based on the current user.